Refactor: move `RecoveryModeDialog` to `dashboard`
