+++ src/modules/admin/components/Profile/ProfileAdvanced.tsx
@@ -8,7 +8,7 @@ import { DialogActionButton } from '~core/Button';
 import Heading from '~core/Heading';
 import ExternalLink from '~core/ExternalLink';
 import NetworkContractUpgradeDialog from '~dashboard/NetworkContractUpgradeDialog';
-import RecoveryModeDialog from '~admin/RecoveryModeDialog';
+import RecoveryModeDialog from '~dashboard/RecoveryModeDialog';
 import {
   Colony,
   useLoggedInUser,
