--- src/modules/admin/components/RecoveryModeDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RecoveryModeDialog';
